+++ src-dbg/flow_storm/debugger/events_processor.clj
@@ -9,17 +9,7 @@
             [flow-storm.debugger.ui.flows.screen :as flows-screen]
             [flow-storm.debugger.ui.flows.code :as flow-code]
             [flow-storm.debugger.ui.utils :as ui-utils]
-            [flow-storm.debugger.ui.state-vars :as ui-vars]
-            [clojure.core.async :as async]
-            [mount.core :refer [defstate]]
-            [flow-storm.utils :as utils]))
-
-(declare start-events-processor)
-(declare stop-events-processor)
-(declare events-processor)
-(defstate events-processor
-  :start (start-events-processor)
-  :stop (stop-events-processor))
+            [flow-storm.debugger.ui.state-vars :as ui-vars]))
 
 (defn- var-instrumented-event [{:keys [var-ns var-name]}]
   (ui-utils/run-later
@@ -63,7 +53,7 @@
 (defn- task-progress-event [{:keys [task-id progress]}]
   (ui-vars/dispatch-task-event :progress task-id progress))
 
-(defn- process-event [[ev-type ev-args-map]]
+(defn process-event [[ev-type ev-args-map]]
   (case ev-type
     :var-instrumented (var-instrumented-event ev-args-map)
     :var-uninstrumented (var-uninstrumented-event ev-args-map)
@@ -74,30 +64,4 @@
     :first-fn-call (first-fn-call-event ev-args-map)
     :tap (tap-event ev-args-map)
     :task-result (task-result-event ev-args-map)
-    :task-progress (task-progress-event ev-args-map)
-    ))
-
-(defn enqueue-event! [e]
-  (async/>!! (:events-chan events-processor) e))
-
-(defn start-events-processor []
-  (utils/log "[Starting Events processor subsystem]")
-  (let [events-chan (async/chan 100)
-        ev-thread (Thread.
-                   (fn []
-                     (try
-                       (loop [ev (async/<!! events-chan)]
-                         (when ev
-                           (process-event ev)
-                           (recur (async/<!! events-chan))))
-                       (catch java.lang.InterruptedException _
-                         (utils/log "Events thread interrupted")))))]
-
-    (.start ev-thread)
-    (utils/log "Events processor started")
-    {:events-chan events-chan
-     :events-thread ev-thread}))
-
-(defn stop-events-processor []
-  (utils/log "[Stopping Events processor subsystem]")
-  (async/close! (:events-chan events-processor)))
+    :task-progress (task-progress-event ev-args-map)))