+++ src-inst/flow_storm/remote_websocket_client.clj
@@ -2,7 +2,7 @@
   (:refer-clojure :exclude [send])
   (:require [flow-storm.json-serializer :as serializer]
             [flow-storm.utils :refer [log log-error] :as utils]
-            )
+            [flow-storm.runtime.events :as rt-events])
   (:import [org.java_websocket.client WebSocketClient]
            [java.net URI]
            [org.java_websocket.handshake ServerHandshake]))
@@ -17,6 +17,16 @@
 (defn remote-connected? []
   (boolean remote-websocket-client))
 
+(defn send [ser-packet]
+  ;; websocket library isn't clear about thread safty of send
+  ;; lets synchronize just in case
+  (locking remote-websocket-client
+    (.send ^WebSocketClient remote-websocket-client ^String ser-packet)))
+
+(defn send-event-to-debugger [ev-packet]
+  (let [ser-packet (serializer/serialize [:event ev-packet])]
+    (send ser-packet)))
+
 (defn start-remote-websocket-client [{:keys [host port on-connected api-call-fn]
                                       :or {host "localhost"
                                            port 7722}}]
@@ -27,6 +37,18 @@
 
                                      (onOpen [^ServerHandshake handshake-data]
                                        (log (format "Connection opened to %s" uri-str))
+
+                                       ;; send all pending events
+                                       (let [pending-events (rt-events/pop-pending-events!)]
+                                         (when (seq pending-events)
+                                           (log "Replaying stored events")
+                                           ;; HACKY: wait a little befor sending the events to be sure
+                                           ;; the debugger that just connected is ready to start
+                                           ;; processing events
+                                           (Thread/sleep 2000)
+                                           (doseq [ev pending-events]
+                                             (send-event-to-debugger ev))))
+
                                        (when on-connected (on-connected)))
 
                                      (onMessage [^String message]
@@ -54,13 +76,3 @@
 
     (alter-var-root #'remote-websocket-client (constantly ws-client))
     ws-client))
-
-(defn send [ser-packet]
-  ;; websocket library isn't clear about thread safty of send
-  ;; lets synchronize just in case
-  (locking remote-websocket-client
-    (.send ^WebSocketClient remote-websocket-client ^String ser-packet)))
-
-(defn send-event-to-debugger [ev-packet]
-  (let [ser-packet (serializer/serialize ev-packet)]
-    (send ser-packet)))