+++ src-inst/flow_storm/remote_websocket_client.cljs
@@ -35,10 +35,10 @@
   (let [ser-packet (serializer/serialize [:event ev-packet])]
     (send ser-packet)))
 
-(defn start-remote-websocket-client [{:keys [host port on-connected api-call-fn]
-                                      :or {host "localhost"
-                                           port 7722}}]
-  (let [uri-str (utils/format "ws://%s:%s/ws" host port)
+(defn start-remote-websocket-client [{:keys [debugger-host on-connected api-call-fn]
+                                      :or {debugger-host "localhost"}}]
+  (let [port 7722
+        uri-str (utils/format "ws://%s:%s/ws" debugger-host port)
         ws-client (web-socket-client-object uri-str)]
 
     (set! (.-onerror ws-client) (fn []