+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 	
 ## New Features
 	
+	- Automatic event retention. For ClojureScript or remote Clojure you don't need to remote-connect by hand
+	  when you want to capture traces before the debugger is connected. Just need to require flow-storm.api on your main.
+	  
 ### Changes
 	
 ### Bugs fixed